ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
langchain==0.2.6
langchain-openai==0.1.14
langchain-community==0.2.6
langchain-chroma==0.1.2
openai==1.35.10
chromadb==0.5.3
pymupdf==1.24.5
python-docx==1.1.2
tiktoken==0.7.0
pydantic==2.8.0
httpx==0.27.0
pymongo==4.8.0
motor==3.5.0
python-multipart==0.0.9
